requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.2
Werkzeug==2.0.3
SQLAlchemy==1.4.32
